requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.27
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.29.0
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
msrest==0.6.21
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
